Wander implemented. It could be improved but works well enough.
There is a bug in the code where Vehicles will be targeting the same food and depending on who gets it, some times, intermittently, the Vehicle/s that didn't get it gets "stuck" on the point where the food used to be. even more rarely, there have been times where Vehicles get attracted to a certain point where food may have been (assumed), and get "stuck" on that point.